requirements.txt
asgiref==3.3.1
boto3==1.17.104
dj-database-url==0.5.0
Django==3.1.6
django-on-heroku==1.1.2
django-storages==1.11.1
gunicorn==20.1.0
h5py==3.1.0
Jinja2==3.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
tensorflow-probability==0.8.0rc0
whitenoise==5.2.0
yfinance==0.1.62